Refactor CloudWeb to use run.py and separate concerns
