+++ cloudweb/__init__.py
@@ -1,0 +1,2 @@
+"""CloudWeb package."""
+
